+++ app/data/generators/assessment-only-details.js
