+++ app/views/_includes/forms/programme-details.html
@@ -1,0 +1,3 @@
+{% if record.route == 'Assessment Only' %}
+  {% include "_includes/forms/programme-details/assessment-only.html" %}
+{% endif %}
