+++ app/views/_includes/summary-cards/programme-details.html
@@ -1,0 +1,3 @@
+{% if record.route == 'Assessment Only' %}
+  {% include "_includes/summary-cards/programme-details/assessment-only.html" %}
+{% endif %}
