+++ app/views/new-record/check-record.html
@@ -64,8 +64,8 @@
 <h3 class="govuk-heading-s">{{ "Degree" | pluralise(record.degree.items | length)}}</h3>
 {% include "_includes/summary-cards/degree-details.html" %}
 
-<h2 class="govuk-heading-m">Assessment details</h2>
-{% include "_includes/summary-cards/assessment-details.html" %}
+<h2 class="govuk-heading-m">Programme details</h2>
+{% include "_includes/summary-cards/programme-details.html" %}
 
 
 {{ govukButton({
