+++ app/views/new-record/programme-details.html
@@ -1,10 +1,10 @@
 {% extends "_templates/_new-record.html" %}
 
-{% set pageHeading = "Assessment details" %}
+{% set pageHeading = "Programme details" %}
 
 {# {% set formAction = "./assessment-details/confirm" %} #}
 
 {% block formContent %}
-  {% include "_includes/forms/assessment-details.html" %}
+  {% include "_includes/forms/programme-details.html" %}
 {% endblock %}
