+++ app/views/record.html
@@ -40,8 +40,8 @@
 
     <h2 class="govuk-heading-m">Trainee record</h2>
     {% include "_includes/summary-cards/student-record.html" %}
-    <h2 class="govuk-heading-m">Assessment details</h2>
-    {% include "_includes/summary-cards/assessment-details.html" %}
+    <h2 class="govuk-heading-m">Programme details</h2>
+    {% include "_includes/summary-cards/programme-details.html" %}
 
   </div>
 <div>
