+++ app/views/record/programme-details/confirm.html
@@ -1,12 +1,12 @@
 {% extends "_templates/_record-form.html" %}
 
-{% set pageHeading = "Confirm assessment details" %}
+{% set pageHeading = "Confirm programme details" %}
 
 {% set formAction = "./update" %}
 
 {% block formContent %}
   <h1 class="govuk-heading-l">{{pageHeading}}</h1>
-  {% include "_includes/summary-cards/assessment-details.html" %}
+  {% include "_includes/summary-cards/programme-details.html" %}
 
   {# {{ govukCheckboxes({
   idPrefix: "waste",
